separate filter logic into its own module
